forge: Clean up the mod meta for platforms
- Actually incorporate deps for the platforms. Minecraft cannot be a dep
for Sponge as this causes classloading issues for all of common on the
ACL
- Minor packaging to consolidate where things are at
- Call out to have services load (does nothing currently, we are not
firing events)
- Minor cleanup
