+++ forge/build.gradle.kts
@@ -35,7 +35,7 @@ loom {
         configureEach {
             if (JavaVersion.current().isJava11Compatible) {
                 vmArgs(
-                        "-Dfabric.log.level=debug",
+                        //"-Dfabric.log.level=debug",
                         "--add-exports=java.base/sun.security.util=ALL-UNNAMED", // ModLauncher
                         "--add-opens=java.base/java.util.jar=ALL-UNNAMED" // ModLauncher
                 )
