+++ forge/src/applaunch/resources/META-INF/services/net.minecraftforge.forgespi.language.IModLanguageProvider
@@ -1,2 +1,2 @@
-org.spongepowered.forge.applaunch.loading.provider.SpongeAPIModLanguageProvider
-org.spongepowered.forge.applaunch.loading.provider.SpongeModLanguageProvider
+org.spongepowered.forge.applaunch.loading.moddiscovery.provider.SpongeAPIModLanguageProvider
+org.spongepowered.forge.applaunch.loading.moddiscovery.provider.SpongeModLanguageProvider
